mods

+++ E06/1.c
@@ -11,7 +11,7 @@ int main(int argc, char **argv){
 
   srand(time(0));
 
-  int a[ASIZE];
+  unsigned short a[ASIZE];
 
   for(int i = 0; i < ASIZE; ++i)
   {
@@ -20,7 +20,7 @@ int main(int argc, char **argv){
 
   for(int i = 0; i < ASIZE; ++i)
   {
-    printf("L'elemento di indice #%d vale %d\n", i, a[i]);
+    printf("L'elemento di indice #%d vale %u\n", i, a[i]);
   }
   return 0;
 }
